New r.fill.category module
r.fill.category substitutes the values of pixels of a given category with values of the surrounding pixels and stores the output in a new raster map. Pixels with other categories are left untouched.
It combines r.neighbors and r.mapcalc. Optionally, it uses r.out.mpeg to create animations.
See the html manual page for an example application on the North Carolina sample dataset.